clippy: clear the 6 remaining lints flagged by -D warnings
Step #7 of the handover review's repair order ("Only then clean
clippy/style fallout") — the runtime issues are resolved, so this
commit clears the residual lint backlog so CI can be tightened to
`cargo clippy --workspace --all-targets -- -D warnings`.

* `tracking::yolox` — drop the inner `#![cfg(feature = "inference")]`.
  The same gate already lives on the `pub mod yolox;` declaration in
  `tracking::mod.rs`, and clippy 1.93's `duplicated_attributes` lint
  flags the pair. Single-source the gate at the mod declaration.

* `tracking::rtmw3d` (YOLOX crop) — `pad_and_clamp_bbox` already
  returns `(u32, u32, u32, u32)`, so the `(cy2 - cy1) as u32` casts
  on `cw` / `ch` were no-ops. Drop them.

* `gui::mod` — the loop over fallback face names was using
  `available.iter().any(|n| *n == name)` where `available: &[&str]`.
  `[T]::contains(&T)` is the canonical form and what clippy
  recommends.

* `avatar::pose_solver` — the per-axis One-Euro vel/pos updates were
  hand-indexed `for i in 0..3 { self.vel[i] += ... }`. Convert to
  `iter_mut().zip(raw_vel.iter())` so clippy's `needless_range_loop`
  is happy. Behaviour is identical.

`cargo clippy --workspace --all-targets -- -D warnings` now passes
clean. `cargo test --workspace --lib` still 118 passed / 0 failed.

+++ src/tracking/rtmw3d.rs
@@ -452,8 +452,8 @@ impl Rtmw3dInference {
                     // / cross-step poses where the leading limb fell
                     // outside the crop.
                     let (cx1, cy1, cx2, cy2) = pad_and_clamp_bbox(&bbox, width, height, 0.25);
-                    let cw = (cx2 - cx1) as u32;
-                    let ch = (cy2 - cy1) as u32;
+                    let cw = cx2 - cx1;
+                    let ch = cy2 - cy1;
                     debug!(
                         "RTMW3D: YOLOX bbox=[{:.0},{:.0},{:.0},{:.0}] score={:.2} → crop {}x{} (orig {}x{})",
                         bbox.x1, bbox.y1, bbox.x2, bbox.y2, bbox.score, cw, ch, width, height
